+++ packages/heroui/eslint.config.ts
@@ -4,30 +4,32 @@ import reactDomPlugin from "eslint-plugin-react-dom";
 import reactHooks from "eslint-plugin-react-hooks";
 import reactRefresh from "eslint-plugin-react-refresh";
 import reactPlugin from "eslint-plugin-react-x";
-import { globalIgnores } from "eslint/config";
+import { defineConfig, globalIgnores } from "eslint/config";
 import tseslint from "typescript-eslint";
 
-export default [
+export default defineConfig([
   globalIgnores(["coverage", "dist", "**/*.d.ts"]),
   ...tseslint.configs.recommended,
   reactPlugin.configs["recommended-typescript"],
   reactDomPlugin.configs.recommended,
-  reactHooks.configs["recommended-latest"],
   reactRefresh.configs.vite,
   prettierConfig,
   {
+    extends: ["react-hooks/recommended"],
     files: ["**/*.{ts,tsx}"],
     plugins: {
       prettier: prettierPlugin,
+      "react-hooks": reactHooks,
     },
     languageOptions: {
       ecmaVersion: 2020,
       parserOptions: {
-        tsconfigdir: process.cwd(),
+        projectService: true,
+        tsconfigdir: import.meta.dirname,
       },
     },
     rules: {
       "prettier/prettier": "error",
     },
   },
-];
+]);